+++ parsnip/parsnip.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2025, The Regents of the University of Michigan
+# Copyright (c) 2025-2026, The Regents of the University of Michigan
 # This file is from the parsnip project, released under the BSD 3-Clause License.
 
 r"""An interface for reading `CIF`_ files in Python.
@@ -100,6 +100,7 @@ from parsnip.patterns import (
     _flatten_or_none,
     _is_data,
     _is_key,
+    _lookup_symops,
     _matrix_from_lengths_and_angles,
     _safe_eval,
     _strip_comments,
@@ -541,9 +542,9 @@ class CifFile:
 
     def build_unit_cell(
         self,
-        n_decimal_places: int = 4,
+        n_decimal_places: int = 3,
         additional_columns: str | Iterable[str] | None = None,
-        parse_mode: Literal["python_float", "sympy"] = "python_float",
+        parse_mode: Literal["rational", "python_float", "sympy"] = "rational",
         verbose: bool = False,
     ):
         """Reconstruct fractional atomic positions from Wyckoff sites and symops.
@@ -560,8 +561,17 @@ class CifFile:
             If the parsed unit cell has more atoms than expected, decrease
             ``n_decimal_places`` to account for noise. If the unit cell has fewer atoms
             than expected, increase ``n_decimal_places`` to ensure atoms are compared
-            with sufficient precision. In many cases, setting ``parse_mode='sympy'``
-            can improve the accuracy of reconstructed unit cells.
+            with sufficient precision.
+
+        .. tip::
+
+            For improved performance, install the `cfractions`_ library. ``parsnip``
+            will automatically detect if it is installed and use the more optimized path
+            when constructing unit cells.
+
+
+        .. _cfractions: https://pypi.org/project/cfractions/
+
 
         Example
         -------
@@ -602,10 +612,12 @@ class CifFile:
                 the Wyckoff site positions. This data is replicated alongside the atomic
                 coordinates and returned in an auxiliary array.
                 Default value = ``None``
-            parse_mode : {'sympy', 'python_float'}, optional
-                Whether to parse lattice sites symbolically (``parse_mode='sympy'``) or
-                numerically (``parse_mode='python_float'``). Sympy is typically more
-                accurate, but may be slower. Default value = ``'python_float'``
+            parse_mode : {'rational', 'sympy', 'python_float'}, optional
+                Whether to parse lattice sites using rational
+                (``parse_mode='rational'``) or floating-point
+                (``parse_mode='python_float'``) arithmetic. 'rational' is more accurate
+                than 'python_float', but may take more time.
+                Default value = ``'rational'``
             verbose : bool, optional
                 Whether to print debug information about the uniqueness checks.
                 Default value = ``False``
@@ -630,7 +642,15 @@ class CifFile:
                 "Sympy is not available! Please set parse_mode='python_float' "
                 "or install sympy."
             )
-        valid_modes = {"sympy", "python_float"}
+        if parse_mode == "sympy":
+            msg = (
+                "The `sympy` parse mode is deprecated in favor of `rational`, the new "
+                "default. `rational` is as accurate as sympy, much faster, and does not"
+                " require any dependencies. Please convert existing codes to use the "
+                "`parse_mode='rational'`."
+            )
+            warnings.warn(msg, category=DeprecationWarning, stacklevel=2)
+        valid_modes = {"rational", "sympy", "python_float"}
         if parse_mode not in valid_modes:
             raise ValueError(f"Parse mode '{parse_mode}' not in {valid_modes}.")
 
@@ -663,11 +683,16 @@ class CifFile:
         )
 
         frac_strs = self._read_wyckoff_positions()
+        if len(frac_strs) == 0:
+            msg = (
+                "No Wyckoff positions were found when constructing unit cell. "
+                f"Found wyckoff_keys: {self._raw_wyckoff_keys or None}"
+            )
+            raise ParseError(msg)
 
         all_frac_positions = [
             _safe_eval(symops_str, *xyz, parse_mode=parse_mode) for xyz in frac_strs
         ]  # Compute N_symmetry_elements coordinates for each Wyckoff site
-
         pos = np.vstack(all_frac_positions)
 
         # Wrap into box - works generally because these are fractional coordinates
@@ -819,7 +844,7 @@ class CifFile:
             if symops is not None:
                 self._symops_key = self._wildcard_mapping[key]
                 return symops
-        return None
+        return _lookup_symops(self)
 
     @property
     def _cell_keys(self):
@@ -844,8 +869,15 @@ class CifFile:
         wyckoff_position_data = [
             self.get_from_loops(key) for key in self.__class__._WYCKOFF_KEYS
         ]
+        wyckoff_position_data = [
+            (col[(col != ".") & (col != "?")][:, None] if col is not None else col)
+            for col in wyckoff_position_data
+        ]
 
-        if all(x is None for x in wyckoff_position_data) and self._strict:
+        if (
+            all((x is None or len(x) == 0) for x in wyckoff_position_data)
+            and self._strict
+        ):
             msg = "No wyckoff position data was found!"
             raise ParseError(msg)
 
@@ -1100,6 +1132,7 @@ class CifFile:
 
                 if n_elements % n_cols != 0:
                     msg = (
+                        f"CifFile('{self._fn}') : "
                         f"Parsed data for table {len(self.loops) + 1} cannot be"
                         f" resolved into a table of the expected size and will be "
                         f"ignored. \nGot n={n_elements} items, which cannot be "
@@ -1183,11 +1216,17 @@ class CifFile:
         # Matcher for <LoopBody> syntactic units, which may span multiple lines.
         # Note that this allows for backtracking, as <SingleQuotedString> and
         # <DoubleQuotedString> values may contain whitespace and quotation marks.
+        # Note that we match in descending order of specificity -- so we first look for
+        # <CharString>, followed by specific <XString> types, then finally ordinary
+        # space-delimited values. This lets use circumvent regular expressions inability
+        # to balance quotes: we simply greedily consume balanced strings where possible,
+        # or fall back to unbalanced tokens where necessary.
         "space_delimited_data": (
             "("
             r";[^;]*?;|"  # Non-semicolon data bracketed by semicolons
-            r"'(?:'\S|[^'])*'|"  # Data with single quotes not followed by \s
-            rf"[^';\"\s]{_PROG_STAR}"  # Additional non-bracketed data
+            r"'(?:\\'|'\S|[^'])*'|"  # Data with single quotes, allowing escapes
+            # Additional non-bracketed data with at most one single or double quote
+            rf"[^';\"\s]{_PROG_STAR}(?:['\"][^';\"\s]{_PROG_STAR})*"
             ")"
         ),
         # Matcher for <Comments> syntactic units